Compose and send mail with -addr flag
If Notmuch is started with the -addr flag, instead of executing a default query, it will open an envelope in compose mode with the To: header already set to the passed address. No From: address is chosen - I instead use an `xdotool` script to type one out.

When a message is successfully sent, the compose window will close automatically to indicate this success, like in `alot`.