Pass arg to Get to refine query
Passing an argument to Get now replaces the current query with a new
query.

By adding query parameters to the quoted text for 'Results of query
"..."', double-clicking to select the entire thing, and using the 2-1
chord to pass that selection to Get, one can quickly refine a query
without opening several new Notmuch windows.